+++ hierarchical_sales_forecasting/b2b_revenue_forecasting/__init__.py
@@ -18,9 +18,9 @@ from b2b_revenue_forecasting.batch import (
     reconcile,
     adjust_many,
 )
-from b2b_revenue_forecasting.pins import Pin, apply_pins, redistribute, concentrate, enforce_identities
+from b2b_revenue_forecasting.pins import Pin, apply_pins, redistribute, concentrate, enforce_identities, reallocate, resplit_by_metric
 
-__version__ = "0.35.0"
+__version__ = "0.36.0"
 
 __all__ = [
     "SalesHierarchy",
@@ -45,5 +45,7 @@ __all__ = [
     "redistribute",
     "concentrate",
     "enforce_identities",
+    "reallocate",
+    "resplit_by_metric",
     "__version__",
 ]